serialize_pythonjob_data when to_dict for PythonJob task
